Chat: moved the contexts in their own subdir.
