+++ src/App.tsx
@@ -3,13 +3,13 @@ import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 import { v1 } from 'uuid';
 
-import { KabelwerkProvider } from './chat/KabelwerkContext';
+import { KabelwerkProvider } from './chat/contexts/KabelwerkContext';
+import { NotificationsProvider } from './chat/contexts/NotificationsContext';
 import Layout from './components/Layout';
 import Placeholder from './components/Placeholder';
 import withProviders from './components/withProviders';
 import routes from './routes';
 import PrivateRoute from './routes/PrivateRoute';
-import { NotificationsProvider } from './chat/NotificationsContext';
 
 export const App = () => (
   <>
