+++ src/chat/contexts/NotificationsContext.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react';
 import { Snackbar } from '@material-ui/core';
 import { Alert } from '@material-ui/lab';
-import { Message } from './interfaces';
+
+import { Message } from '../interfaces';
 
 enum Severity {
   Info = 'info',
