+++ src/chat/hooks/useNotification.ts
@@ -1,4 +1,4 @@
-import { NotificationsContext } from '../NotificationsContext';
+import { NotificationsContext } from '../contexts/NotificationsContext';
 import * as React from 'react';
 
 const useNotification = () => {
